backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
qdrant-client==1.12.1
langchain==0.3.13
langchain-google-genai==2.0.8
langchain-community==0.3.13
google-generativeai==0.8.3
httpx==0.28.1
python-multipart==0.0.19
